requirements.txt
alembic==1.13.1
amqp==5.2.0
apispec==6.6.1
bcrypt==4.1.3
billiard==4.2.0
blinker==1.8.2
boto3==1.34.124
botocore==1.34.124
cachelib==0.9.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
flask==3.0.3
flask-apispec==0.11.4
flask-bcrypt==1.0.1
flask-caching==2.3.0
flask-cors==4.0.1
flask-mail==0.10.0
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
mako==1.3.5
markupsafe==2.1.5
marshmallow==3.21.3
packaging==24.1
prompt-toolkit==3.0.47
psycopg2==2.9.9
pyjwt==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.5
s3transfer==0.10.1
six==1.16.0
sqlalchemy==2.0.30
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webargs==8.4.0
werkzeug==3.0.3
